add base dashboard controller test

+++ console/src/main/java/com/mvc/cryptovault/console/dashboard/controller/DAppProjectController.java
@@ -27,6 +27,9 @@ public class DAppProjectController extends BaseController {
     @DeleteMapping("{id}")
     public Result<Boolean> deleteProject(@PathVariable("id") BigInteger id) {
         var appProject = appProjectService.findById(id);
+        if(null == appProject){
+            return new Result<>(true);
+        }
         //只有未展示的项目才可以删除
         Assert.isTrue(null != appProject && appProject.getVisiable() == 0, "项目已展示,无法删除");
         appProjectService.deleteById(id);
